Moved entity, repository, and translators over to model repo
