+++ src/main/java/com/rackspace/salus/monitor_management/config/DatabaseConfig.java
@@ -20,7 +20,7 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@EntityScan({"com.rackspace.salus.telemetry.entities", "com.rackspace.salus.monitor_management.entities"})
+@EntityScan("com.rackspace.salus.telemetry.entities")
 @EnableJpaRepositories("com.rackspace.salus.telemetry.repositories")
 @Configuration
 public class DatabaseConfig {
