--- src/main/java/com/rackspace/salus/monitor_management/entities/package-info.java
@@ -1,22 +1,0 @@
-/*
- * Copyright 2019 Rackspace US, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * This package contains entities that are local to Monitor Management and unlike the
- * <code>com.rackspace.salus.telemetry.entities</code> package in
- * <code>salus-telemetry-model</code>, these entities are "not shared" with other modules.
- */
-package com.rackspace.salus.monitor_management.entities;
