+++ src/main/java/com/rackspace/salus/monitor_management/services/MonitorContentTranslationService.java
@@ -20,14 +20,14 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.github.zafarkhaja.semver.Version;
-import com.rackspace.salus.monitor_management.entities.MonitorTranslationOperator;
-import com.rackspace.salus.monitor_management.repositories.MonitorTranslationOperatorRepository;
 import com.rackspace.salus.monitor_management.web.model.BoundMonitorDTO;
 import com.rackspace.salus.monitor_management.web.model.MonitorTranslationOperatorCreate;
-import com.rackspace.salus.monitor_management.web.model.translators.MonitorTranslator;
 import com.rackspace.salus.telemetry.entities.BoundMonitor;
+import com.rackspace.salus.telemetry.entities.MonitorTranslationOperator;
 import com.rackspace.salus.telemetry.model.AgentType;
 import com.rackspace.salus.telemetry.model.NotFoundException;
+import com.rackspace.salus.telemetry.repositories.MonitorTranslationOperatorRepository;
+import com.rackspace.salus.telemetry.translators.MonitorTranslator;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
