+++ src/test/java/com/rackspace/salus/monitor_management/services/MonitorContentTranslationServiceTest.java
@@ -24,16 +24,16 @@ import static org.mockito.Mockito.when;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.rackspace.salus.monitor_management.entities.MonitorTranslationOperator;
-import com.rackspace.salus.monitor_management.repositories.MonitorTranslationOperatorRepository;
 import com.rackspace.salus.monitor_management.web.model.BoundMonitorDTO;
-import com.rackspace.salus.monitor_management.web.model.translators.MonitorTranslator;
-import com.rackspace.salus.monitor_management.web.model.translators.RenameFieldTranslator;
 import com.rackspace.salus.telemetry.entities.BoundMonitor;
 import com.rackspace.salus.telemetry.entities.Monitor;
+import com.rackspace.salus.telemetry.entities.MonitorTranslationOperator;
 import com.rackspace.salus.telemetry.model.AgentType;
 import com.rackspace.salus.telemetry.model.ConfigSelectorScope;
 import com.rackspace.salus.telemetry.model.MonitorType;
+import com.rackspace.salus.telemetry.repositories.MonitorTranslationOperatorRepository;
+import com.rackspace.salus.telemetry.translators.MonitorTranslator;
+import com.rackspace.salus.telemetry.translators.RenameFieldTranslator;
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.List;
